Use replacement pattern for case-insensitive highlights
This commit improves highlighting by using a replacement pattern as described here:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/replace#Specifying_a_string_as_a_parameter

When the user searches using a string that includes an uppercase character, ${this.value} will insert in the highlights - whether the matched substring had a capital in it or not. Using the $&-pattern fixes this, as it will insert the originally matched (sub)string.