Project clarification needed
I think there should at least be a comment in the start index-START.html that clarifies what we are supposed to achieve with this exercise. Your description at the beginning of the video does not match what we are supposed to do. This is the comment I have added to this file:

<!-- project clarification: sorting band names strings in band array IGNORING the any A, An, or The articles at the beginning of each band name string, NOT WITHOUT ARTICLES, MEANING not removing them -->

I think this would be of help to many in finding the solution on their own. I was very confused because I thought I was supposed to remove the articles entirely. Thanks!