🥩 Replace reference to missing image with a steak
Okay okay, so you don't have to use this picture of a steak my
girlfriend cooked the other night (but if you want to, please be my
guest, it was really tasty) -- but the original image that was
referenced here is no longer there. I couldn't easily find where that
image lived, so I just replaced it with a photo of a steak.