11 Custom Video Player fix mouseup event
Make mouseup event on document instead of progress so that the mousedown flag will be reset even if the mouse is released somewhere other than the progress bar. As it stands now if you click drag on the progress bar, move off the bar and release, then mouse down is still activated and scrub will be called whenever the mouse runs over the progress bar.