Detect intermediate input changes with a better event -> the input event
NOTE: at newer version of Firefox, Chrome and Edge, the mousemove event do not cover intermediate changes on the color input … but this commit with the new input event works fine, maybe a simple suggestion on how to improve that. Thank you for your cool course, it’s really good and helpful.