Master <!-- 👋👋👋👋👋👋👋👋👋👋👋👋👋👋 👋👋👋Hello Friend!👋👋👋👋 👋👋👋👋👋👋👋👋👋👋👋👋👋👋
Thanks for Submitting a pull request. Before you hit that button please make sure:

These files are meant to be 1:1 copies of what is done in the video. If you found a better / different way to do things or fixed a small bug, that is great great, but I will be keeping them the same as the videos to avoid confusing.

Spelling mistakes / CSS updates / other clarifications are welcome as long as they don't change what is shown in the videos.

I encourage you to blog about your implementation and add the link to this repo - that way everyone can benefit from it.

-->